tests/wpt/tests/IndexedDB/idbfactory_cmp3.htm

<!DOCTYPE html>
<meta charset=utf-8>
<title>IDBFactory.cmp() - compared keys in different types</title>
<link rel="author" title="Mozilla" href="https://www.mozilla.org">
<link rel="help" href="http://w3c.github.io/IndexedDB/#key-construct">
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>

<script>
  test(function() {
    assert_equals(indexedDB.cmp([0], new Uint8Array([0])), 1, "Array > Binary");
  }, "Array v.s. Binary");

  test(function() {
    assert_equals(indexedDB.cmp(new Uint8Array([0]), "0"), 1, "Binary > String");
  }, "Binary v.s. String");

  test(function() {
    assert_equals(indexedDB.cmp("", new Date(0)), 1, "String > Date");
  }, "String v.s. Date");

  test(function() {
    assert_equals(indexedDB.cmp(new Date(0), 0), 1, "Date > Number");
  }, "Date v.s. Number");
</script>

<div id=log></div>
